A smal update!

+++ UnZipMain.py
@@ -32,7 +32,8 @@ while(True):
                             ctypes.windll.user32.MessageBoxW(0, "The program didn't find a matching password for file:\n "+zipfi+"\nPlease add the password to the config.json file!", "Error: Password not found!", 0)
                         cr = 0
             if cr:
-                os.remove(zipfi)
+                if config["rmfile"]:
+                    os.remove(zipfi)
             cr = 1
     lisofi = []
     sleep(config["update-time"]*60)
